Fix Windows portability bugs in C extensions; add cibuildwheel CI
Real bugs, not just missing wheels -- the source wouldn't even build on
Windows before this, and one bug would have silently corrupted results:

- c_entropy.c: replaced two C99 VLAs (double hist[n_hist], size_t
  ns[k-1]) with malloc/free -- MSVC has never supported VLAs, at any
  version. Also fixed a pre-existing off-by-one buffer overflow in ns
  (sized k-1 but written up to index k-1, needs k elements).
- _c_entropy_build.py / _se_build.py: libraries=["m"] (POSIX libm) and
  extra_compile_args=["-O3"] are Unix/GCC-only; made conditional on
  sys.platform (no separate libm under MSVC; /O2 instead of -O3).
- se.c: sample_entropy_disc(_fast) took `long*`, but the Python side
  always passes np.int64 (8 bytes). long is 8 bytes on Linux/macOS but
  only 4 bytes on Windows (LLP64) -- silently reinterprets/corrupts the
  data instead of erroring. Changed to `long long` (64-bit on every
  platform) throughout the C signatures, the CFFI .cdef, and the
  mstsa.py cast() call. Also fixed `long N = n*(n-1)/2` overflowing a
  32-bit Windows long for sequences longer than ~65k samples (this
  project's own EEG data is 300k+ samples).
- _lz76_build.py: same -O3/MSVC fix for consistency.

Verified locally: rebuilt all three extensions, hk/lz76/sample_entropy_disc
match the pure-Python fallback exactly, full test suite (32 tests) passing.

.github/workflows/wheels.yml: cibuildwheel-based CI to build real
precompiled wheels (Linux manylinux, Windows win_amd64, macOS
Intel+Apple Silicon; CPython 3.9-3.13) and publish to PyPI, so
`pip install mstsa` needs no compiler on any platform -- needed for
zero-friction install at an upcoming workshop. Triggers on `v*` tags
(auto-publish) or manual workflow_dispatch (optional dry-run without
publishing). Requires a PYPI_API_TOKEN repo secret to actually publish.

Also untracked mstsa/_se.c (a CFFI-auto-generated wrapper file that had
been accidentally committed; its siblings _c_entropy.c/_lz76.c were
correctly never tracked) and added the general pattern to .gitignore.

+++ mstsa/src/se.c
@@ -41,7 +41,7 @@ double sample_entropy_cont(double* x, size_t n, size_t m, size_t tau, double r) 
     return se;
 }
 
-double sample_entropy_disc(long* x, size_t n, size_t m, size_t tau) {
+double sample_entropy_disc(long long* x, size_t n, size_t m, size_t tau) {
     size_t i, j, k;
     double A, B, se;
     B = 0.0;
@@ -73,7 +73,9 @@ int sample_entropy_cont_fast(double* se, double* x, size_t n, size_t m, double r
     size_t i, j, k;
     size_t nj, jj, m1;
     double x1;
-    long N = (long) (n*(n-1)/2);
+    // n*(n-1)/2 overflows a 32-bit long (Windows) for realistic sequence
+    // lengths (e.g. n > ~65536); long long is 64-bit on every platform.
+    long long N = (long long) (n*(n-1)/2);
     double* A = (double *) calloc(m, sizeof(double));
     double* B = (double *) calloc(m, sizeof(double));
     double* p = (double *) calloc(m, sizeof(double));
@@ -121,11 +123,13 @@ int sample_entropy_cont_fast(double* se, double* x, size_t n, size_t m, double r
     return 0;
 }
 
-int sample_entropy_disc_fast(double* se, long* x, size_t n, size_t m) {
+int sample_entropy_disc_fast(double* se, long long* x, size_t n, size_t m) {
     size_t i, j, k;
     size_t nj, jj, m1;
-    long x1;
-    long N = (long) (n*(n-1)/2);
+    long long x1;
+    // n*(n-1)/2 overflows a 32-bit long (Windows) for realistic sequence
+    // lengths (e.g. n > ~65536); long long is 64-bit on every platform.
+    long long N = (long long) (n*(n-1)/2);
     double* A = (double *) calloc(m, sizeof(double));
     double* B = (double *) calloc(m, sizeof(double));
     double* p = (double *) calloc(m, sizeof(double));
